+++ src/collective/pivot/utils.py
@@ -1,12 +1,10 @@
 # -*- coding: utf-8 -*-
 from plone import api
-from zope.component import queryUtility
-from zope.i18n.interfaces import ITranslationDomain
+from zope.i18n import translate
 
 
 def _(msgid, context, domain="collective.pivot", mapping=None):
-    translation_domain = queryUtility(ITranslationDomain, domain)
-    return translation_domain.translate(msgid, context=context.REQUEST, mapping=mapping)
+    return translate(msgid, context=context.REQUEST, mapping=mapping)
 
 
 def add_family(context, family_id, title):